Simplify comment rating API
You can now only set or get the comment rating with POST or GET
respectively. To remove a rating, set the rating to 0. Also, the API now
assumes the use of cookies such that a client is identified by the ID in
the cookie (to be implemented). That's why no rating-id is passed and
'ratings' have been changed to the singular 'rating'.